Fixes to refactoring
